+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/model/AnalysisResult.java
@@ -46,7 +46,7 @@ public class AnalysisResult extends HasId {
     // Reference to the QPU instance which is stored in the QProv database
     private String qpu;
     private String provider;
-    private String sdkConnector;
+    private String compiler;
 
     @ManyToOne
     private Implementation implementation;
