+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/web/controller/RootController.java
@@ -31,17 +31,14 @@ import java.util.stream.Collectors;
 
 import org.planqk.nisq.analyzer.core.Constants;
 import org.planqk.nisq.analyzer.core.control.NisqAnalyzerControlService;
-import org.planqk.nisq.analyzer.core.model.AnalysisResult;
 import org.planqk.nisq.analyzer.core.model.CompilationJob;
 import org.planqk.nisq.analyzer.core.model.ImplementationSelectionJob;
 import org.planqk.nisq.analyzer.core.repository.CompilationJobRepository;
 import org.planqk.nisq.analyzer.core.repository.ImplementationSelectionJobRepository;
 import org.planqk.nisq.analyzer.core.web.Utils;
-import org.planqk.nisq.analyzer.core.web.dtos.entities.AnalysisResultDto;
-import org.planqk.nisq.analyzer.core.web.dtos.entities.AnalysisResultListDto;
 import org.planqk.nisq.analyzer.core.web.dtos.entities.CompilationJobDto;
 import org.planqk.nisq.analyzer.core.web.dtos.entities.CompilerAnalysisResultListDto;
-import org.planqk.nisq.analyzer.core.web.dtos.entities.ImplementationSelectionJobDto;
+import org.planqk.nisq.analyzer.core.web.dtos.entities.AnalysisJobDto;
 import org.planqk.nisq.analyzer.core.web.dtos.entities.ParameterDto;
 import org.planqk.nisq.analyzer.core.web.dtos.entities.ParameterListDto;
 import org.planqk.nisq.analyzer.core.web.dtos.requests.CompilerSelectionDto;
@@ -136,7 +133,7 @@ public class RootController {
     @Operation(responses = {@ApiResponse(responseCode = "200"), @ApiResponse(responseCode = "400", content = @Content),
             @ApiResponse(responseCode = "500", content = @Content)}, description = "Select implementations for an algorithm")
     @PostMapping("/" + Constants.SELECTION)
-    public HttpEntity<ImplementationSelectionJobDto> selectImplementations(@RequestBody SelectionRequestDto params) {
+    public HttpEntity<AnalysisJobDto> selectImplementations(@RequestBody SelectionRequestDto params) {
         LOG.debug("Post to select implementations for algorithm with Id {} received.", params.getAlgorithmId());
 
         if (Objects.isNull(params.getAlgorithmId())) {
@@ -163,7 +160,7 @@ public class RootController {
             return new ResponseEntity("No prolog engine accessible from the server. Selection not possible!", HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
-        ImplementationSelectionJobDto dto = ImplementationSelectionJobDto.Converter.convert(job);
+        AnalysisJobDto dto = AnalysisJobDto.Converter.convert(job);
         dto.add(linkTo(methodOn(AnalysisResultController.class).getImplementationSelectionJob(job.getId())).withSelfRel());
         return new ResponseEntity<>(dto, HttpStatus.OK);
     }
