+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/web/dtos/entities/AnalysisResultDto.java
@@ -40,7 +40,7 @@ public class AnalysisResultDto extends RepresentationModel<AnalysisResultDto> {
 
     String provider;
 
-    String sdkConnector;
+    String compiler;
 
     ImplementationDto implementation;
 
@@ -59,7 +59,7 @@ public class AnalysisResultDto extends RepresentationModel<AnalysisResultDto> {
             dto.setId(object.getId());
             dto.setQpu(object.getQpu());
             dto.setProvider(object.getProvider());
-            dto.setSdkConnector(object.getSdkConnector());
+            dto.setCompiler(object.getCompiler());
             dto.setImplementation(ImplementationDto.Converter.convert(object.getImplementation()));
             dto.setAnalyzedDepth(object.getAnalyzedDepth());
             dto.setAnalyzedWidth(object.getAnalyzedWidth());
